Python Script for Blacklist IP API check
This is a python script that Git Copilot generated when I asked it to cross reference my router log file with known Blacklist IP's. I asked it to use an API from blacklistchecker, but as it could not run the script, it generated this python script for me to run and have it check the IP's from my log file. THe IPs from my log file are hard coded in the script.